requirements.txt
scapy==2.5.0
python-nmap=0.7.1